Cleanup and rearrange src/test, see full log
A lot of tests that aren't yet implemented were removed.

The testing interface was removed. There isn't a good reason to have one at
the moment. All tests will be run as a regular program anyway, there isn't a
need to have one file with a main call a standard function in another class
that implements an interface.

The "all" tests command will be implemented as a simple find target/test and
exec script, most likely. If it is not, you would have to manually code a call
to each test anyway. We can just use the test's main for this.

TestPrimeFactorization has been changed into TestTrialDivison to reflect
changes in src/main